Conflicts:
	src/Battlescape/BattlescapeState.cpp

+++ src/Battlescape/TileEngine.cpp
@@ -422,7 +422,7 @@ bool TileEngine::visible(BattleUnit *currentUnit, Tile *tile)
 		_trajectory.clear();
 		calculateLine(originVoxel, scanVoxel, true, &_trajectory, currentUnit);
 		Tile *t = _save->getTile(currentUnit->getPosition());
-		int visibleDistance = _trajectory.size();
+		size_t visibleDistance = _trajectory.size();
 		for (size_t i = 0; i < _trajectory.size(); i++)
 		{
 			if (t != _save->getTile(Position(_trajectory.at(i).x/16,_trajectory.at(i).y/16, _trajectory.at(i).z/24)))
@@ -2587,8 +2587,9 @@ bool TileEngine::validMeleeRange(BattleUnit *attacker, BattleUnit *target, int d
  * Validates the melee range between a tile and a unit.
  * @param pos Position to check from.
  * @param direction Direction to check.
- * @param size For large units, we have to do extra checks.
+ * @param attacker The attacking unit.
  * @param target The unit we want to attack, 0 for any unit.
+ * @param dest Destination position.
  * @return True when the range is valid.
  */
 bool TileEngine::validMeleeRange(Position pos, int direction, BattleUnit *attacker, BattleUnit *target, Position *dest)
@@ -2644,6 +2645,7 @@ bool TileEngine::validMeleeRange(Position pos, int direction, BattleUnit *attack
 
 /**
  * Gets the AI to look through a window.
+ * @param position Current position.
  * @return Direction or -1 when no window found.
  */
 int TileEngine::faceWindow(const Position &position)
@@ -2668,7 +2670,7 @@ int TileEngine::faceWindow(const Position &position)
  * @param action The action to validate.
  * @param originVoxel The origin point of the action.
  * @param targetVoxel The target point of the action.
- * @param curvature The curvature of the throw.
+ * @param curve The curvature of the throw.
  * @param voxelType The type of voxel at which this parabola terminates.
  * @return Validity of action.
  */
@@ -2678,7 +2680,7 @@ bool TileEngine::validateThrow(BattleAction &action, Position originVoxel, Posit
 	double curvature = 0.5;
 	if (action.type == BA_THROW)
 	{
-		curvature = std::max(0.48, 1.73 / sqrt(sqrt((double)(action.actor->getStats()->strength / action.weapon->getRules()->getWeight()))) + (action.actor->isKneeled()? 0.1 : 0.0));
+		curvature = std::max(0.48, 1.73 / sqrt(sqrt((double)(action.actor->getStats()->strength) / (double)(action.weapon->getRules()->getWeight()))) + (action.actor->isKneeled()? 0.1 : 0.0));
 	}
 	Tile *targetTile = _save->getTile(action.target);
 	// object blocking - can't throw here
@@ -2738,6 +2740,8 @@ void TileEngine::recalculateFOV()
 
 /**
  * Returns the direction from origin to target.
+ * @param origin The origin point of the action.
+ * @param target The target point of the action.
  * @return direction.
  */
 int TileEngine::getDirectionTo(const Position &origin, const Position &target) const
@@ -2784,6 +2788,12 @@ int TileEngine::getDirectionTo(const Position &origin, const Position &target) c
 	return dir;
 }
 
+/**
+ * Gets the origin voxel of a certain action.
+ * @param action Battle action.
+ * @param tile Pointer to the action tile.
+ * @return origin position.
+ */
 Position TileEngine::getOriginVoxel(BattleAction &action, Tile *tile)
 {
 	
@@ -2853,7 +2863,7 @@ Position TileEngine::getOriginVoxel(BattleAction &action, Tile *tile)
 	return originVoxel;
 }
 
-/*
+/**
  * mark a region of the map as "dangerous" for a turn.
  * @param pos is the epicenter of the explosion.
  * @param radius how far to spread out.
